Migrate landing page to functional components
