+++ i18n/en.pot
@@ -5,88 +5,13 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2020-05-01T06:21:13.407Z\n"
-"PO-Revision-Date: 2020-05-01T06:21:13.407Z\n"
+"POT-Creation-Date: 2020-05-04T06:46:09.001Z\n"
+"PO-Revision-Date: 2020-05-04T06:46:09.001Z\n"
 
 msgid "Field {{field}} cannot be blank"
 msgstr ""
 
 msgid "You need to select at least one {{element}}"
-msgstr ""
-
-msgid "Cannot read file"
-msgstr ""
-
-msgid "Object not found in database"
-msgstr ""
-
-msgid "Select a organisation units to import data"
-msgstr ""
-
-msgid "Selected organisation unit is not associated with the dataset"
-msgstr ""
-
-msgid "Imported"
-msgstr ""
-
-msgid "Updated"
-msgstr ""
-
-msgid "Ignored"
-msgstr ""
-
-msgid "Deleted"
-msgstr ""
-
-msgid "Data Set"
-msgstr ""
-
-msgid "Program"
-msgstr ""
-
-msgid "Existing data values"
-msgstr ""
-
-msgid ""
-"There are {{dataValuesSize}} data values in the database for this "
-"organisation unit and periods. If you proceed, all those data values will "
-"be deleted and only the ones in the spreadsheet will be saved. Are you sure?"
-msgstr ""
-
-msgid "Proceed"
-msgstr ""
-
-msgid "Cancel"
-msgstr ""
-
-msgid "Template Generation"
-msgstr ""
-
-msgid "Start Year"
-msgstr ""
-
-msgid "End Year"
-msgstr ""
-
-msgid "No capture organisations units"
-msgstr ""
-
-msgid "Download template"
-msgstr ""
-
-msgid "Bulk Import"
-msgstr ""
-
-msgid "Drag and drop file to import"
-msgstr ""
-
-msgid "data values"
-msgstr ""
-
-msgid "No capture org unit match element org units"
-msgstr ""
-
-msgid "Import data"
 msgstr ""
 
 msgid "Settings saved"
@@ -108,6 +33,12 @@ msgid "User groups for Template Generation"
 msgstr ""
 
 msgid "User groups with access to Settings"
+msgstr ""
+
+msgid "Data Set"
+msgstr ""
+
+msgid "Program"
 msgstr ""
 
 msgid "Custom"
@@ -211,4 +142,73 @@ msgstr ""
 msgid ""
 "Cannot import: Versions do not match (database={{dbVersion}}, "
 "file={{sheetVersion}})"
+msgstr ""
+
+msgid "Cannot read file"
+msgstr ""
+
+msgid "Object not found in database"
+msgstr ""
+
+msgid "Select a organisation units to import data"
+msgstr ""
+
+msgid "Selected organisation unit is not associated with the dataset"
+msgstr ""
+
+msgid "Imported"
+msgstr ""
+
+msgid "Updated"
+msgstr ""
+
+msgid "Ignored"
+msgstr ""
+
+msgid "Deleted"
+msgstr ""
+
+msgid "Existing data values"
+msgstr ""
+
+msgid ""
+"There are {{dataValuesSize}} data values in the database for this "
+"organisation unit and periods. If you proceed, all those data values will "
+"be deleted and only the ones in the spreadsheet will be saved. Are you sure?"
+msgstr ""
+
+msgid "Proceed"
+msgstr ""
+
+msgid "Cancel"
+msgstr ""
+
+msgid "Template Generation"
+msgstr ""
+
+msgid "Start Year"
+msgstr ""
+
+msgid "End Year"
+msgstr ""
+
+msgid "No capture organisations units"
+msgstr ""
+
+msgid "Download template"
+msgstr ""
+
+msgid "Bulk Import"
+msgstr ""
+
+msgid "Drag and drop file to import"
+msgstr ""
+
+msgid "data values"
+msgstr ""
+
+msgid "No capture org unit match element org units"
+msgstr ""
+
+msgid "Import data"
 msgstr ""
