+++ src/webapp/pages/root/Root.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import LandingPage from "../landing-page/LandingPage";
+
+const Root = () => {
+    return <LandingPage />;
+};
+
+export default Root;
